refactor-compact-tiles * Move getTransportHeaderIcons * Delete duplicat getTransportHeaderIcons
