+++ src/dashboards/BusStop/components/BusStopTile/BusStopTile.tsx
@@ -1,41 +1,16 @@
 import React, { useMemo } from 'react'
 import { Table, TableRow, TableHead, HeaderCell, TableBody } from '@entur/table'
-import {
-    StopPlaceWithDepartures,
-    LineData,
-    IconColorType,
-} from '../../../../types'
+import { StopPlaceWithDepartures } from '../../../../types'
 import { useSettings } from '../../../../settings/SettingsProvider'
 import { WalkInfo } from '../../../../logic/use-walk-info/useWalkInfo'
 import { BusStopTableRow } from '../BusStopTableRow/BusStopTableRow'
-import { isNotNullOrUndefined } from '../../../../utils/typeguards'
-import { unique } from '../../../../utils/array'
 import {
-    getIcon,
     getIconColorType,
-    getTransportIconIdentifier,
+    getTransportHeaderIcons,
 } from '../../../../utils/icon'
 import { Tile } from '../../../../components/Tile/Tile'
 import { TileHeader } from '../../../../components/TileHeader/TileHeader'
 import classes from './BusStopTile.module.scss'
-
-function getTransportHeaderIcons(
-    departures: LineData[],
-    iconColorType: IconColorType,
-): JSX.Element[] {
-    const transportModes = unique(
-        departures.map(({ type, subType }) => ({ type, subType })),
-        (a, b) =>
-            getTransportIconIdentifier(a.type, a.subType) ===
-            getTransportIconIdentifier(b.type, b.subType),
-    )
-
-    const transportIcons = transportModes.map(({ type, subType }) => ({
-        icon: getIcon(type, iconColorType, subType),
-    }))
-
-    return transportIcons.map(({ icon }) => icon).filter(isNotNullOrUndefined)
-}
 
 const BusStopTile = ({
     stopPlaceWithDepartures,
